Add contrib module for MIMIC (#110)
MIMIC (Medical Information Mart for Intensive Care) is a large database of deidentified EHR and medical imaging data. For more information, see the MIMIC documentation. This Meerkat integration enables quick exploration the MIMIC database via the DataPanel.

Other changes:

Fix bug where overwritten columns were not removed from blockmanager, now tested for with test_overwrite_column

+++ meerkat/ops/merge.py
@@ -87,7 +87,11 @@ def merge(
         merged_dp.visible_columns = (
             merged_dp.visible_columns[-1:] + merged_dp.visible_columns[:-1]
         )
-    if not keep_indexes:
+    if (
+        not keep_indexes
+        and ("index" + suffixes[0]) in merged_dp
+        and ("index" + suffixes[1]) in merged_dp
+    ):
         merged_dp.remove_column("index" + suffixes[0])
         merged_dp.remove_column("index" + suffixes[1])
